feat: separate devices
