+++ custom_components/venta/devices/100.py
@@ -1,0 +1,106 @@
+"""Venta AS100 setup functions."""
+
+from __future__ import annotations
+
+from homeassistant.components.sensor import SensorDeviceClass, SensorStateClass
+from homeassistant.const import (
+    ATTR_TEMPERATURE,
+    PERCENTAGE,
+    UnitOfTemperature,
+)
+from homeassistant.helpers.entity_platform import AddEntitiesCallback
+
+from custom_components.venta.utils import skip_zeros
+
+from ..const import (
+    ATTR_HUMIDITY,
+    ATTR_TOLUENE,
+    ATTR_VOC,
+)
+from ..venta import VentaDataUpdateCoordinator
+from ..venta_entity import (
+    VentaSensor,
+    VentaSensorEntityDescription,
+)
+
+# TODO:  Check how SensorStatus works e.g. "0000001F"
+
+
+async def async_setup_humidifier(
+    coordinator: VentaDataUpdateCoordinator, async_add_entities: AddEntitiesCallback
+) -> None:
+    """Set up humidifiers for Venta AS100."""
+    pass
+
+
+async def async_setup_binary_sensor(
+    coordinator: VentaDataUpdateCoordinator, async_add_entities: AddEntitiesCallback
+) -> None:
+    """Set up binary sensors for Venta AS100."""
+    pass
+
+
+async def async_setup_sensor(
+    coordinator: VentaDataUpdateCoordinator, async_add_entities: AddEntitiesCallback
+) -> None:
+    """Set up sensors for Venta AS100."""
+    descriptions = [
+        VentaSensorEntityDescription(
+            key=ATTR_TEMPERATURE,
+            translation_key=ATTR_TEMPERATURE,
+            device_class=SensorDeviceClass.TEMPERATURE,
+            state_class=SensorStateClass.MEASUREMENT,
+            native_unit_of_measurement=UnitOfTemperature.CELSIUS,
+            suggested_display_precision=1,
+            value_func=lambda coordinator: coordinator.data.measure.get("Temperature"),
+        ),
+        VentaSensorEntityDescription(
+            key=ATTR_HUMIDITY,
+            translation_key=ATTR_HUMIDITY,
+            state_class=SensorStateClass.MEASUREMENT,
+            device_class=SensorDeviceClass.HUMIDITY,
+            native_unit_of_measurement=PERCENTAGE,
+            suggested_display_precision=1,
+            value_func=lambda coordinator: coordinator.data.measure.get("Humidity"),
+        ),
+        VentaSensorEntityDescription(
+            key=ATTR_VOC,
+            translation_key=ATTR_VOC,
+            device_class=SensorDeviceClass.VOLATILE_ORGANIC_COMPOUNDS_PARTS,
+            state_class=SensorStateClass.MEASUREMENT,
+            value_func=lambda coordinator: skip_zeros(
+                coordinator.data.measure.get("Voc"),
+            ),
+        ),
+        VentaSensorEntityDescription(
+            key=ATTR_TOLUENE,
+            translation_key=ATTR_TOLUENE,
+            device_class=SensorDeviceClass.VOLATILE_ORGANIC_COMPOUNDS_PARTS,
+            state_class=SensorStateClass.MEASUREMENT,
+            value_func=lambda coordinator: coordinator.data.measure.get("Toluene"),
+        ),
+    ]
+    async_add_entities(
+        [VentaSensor(coordinator, description) for description in descriptions]
+    )
+
+
+async def async_setup_switch(
+    coordinator: VentaDataUpdateCoordinator, async_add_entities: AddEntitiesCallback
+) -> None:
+    """Set up switches for Venta AS100."""
+    pass
+
+
+async def async_setup_light(
+    coordinator: VentaDataUpdateCoordinator, async_add_entities: AddEntitiesCallback
+) -> None:
+    """Set up lights for Venta AS100."""
+    pass
+
+
+async def async_setup_select(
+    coordinator: VentaDataUpdateCoordinator, async_add_entities: AddEntitiesCallback
+) -> None:
+    """Set up selects for Venta AS100."""
+    pass
